Converted the UTXO double-buffer to a MyISAM table.
Previously, the unspent_transaction_outputs_buffer table used the MEMORY engine;
this table's purpose is to serve as a write-only buffer when committing UTXOs from
the cache to disk.  However, using the MEMORY engine for this buffer, while faster,
reduces the maximum number of UTXOs within the actual utxo cache. Setting this table
as MyISAM sacrifices a small amount of performance during UTXO commit for better
overall UTXO caching performance.